Some more refactoring of the boyar driver and data purge tests
