+++ test/helpers/eventually.go
@@ -1,0 +1,15 @@
+package helpers
+
+import "time"
+
+const eventuallyIterations = 50
+
+func Eventually(timeout time.Duration, f func() bool) bool {
+	for i := 0; i < eventuallyIterations; i++ {
+		if f() {
+			return true
+		}
+		time.Sleep(timeout / eventuallyIterations)
+	}
+	return false
+}
